predict_service/requirements.txt
grpcio
grpcio-tools
protobuf
pandas
numpy
scikit-learn
joblib
